# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Zed CV API
  description: AI-powered job matching platform for Zambia
  version: 0.1.0
  contact:
    email: convergeozambia@gmail.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.zedcv.com/api/v1
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AdminMatchRow:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        user_phone: { type: string }
        job_id: { type: string, format: uuid }
        job_title: { type: string }
        job_company: { type: string }
        score: { type: number }
        status: { type: string }
        created_at: { type: string, format: date-time }
    AdminMatchList:
      type: object
      properties:
        matches: { type: array, items: { $ref: '#/components/schemas/AdminMatchRow' } }
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
        pages: { type: integer }
    AdminTierBreakdown:
      type: object
      properties:
        free: { type: integer }
        starter: { type: integer }
        professional: { type: integer }
        super_standard: { type: integer }
        total_active: { type: integer }
    AdminSubscriptionRow:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        user_phone: { type: string }
        full_name: { type: string }
        tier: { type: string, enum: [free, starter, professional, super_standard] }
        status: { type: string, enum: [active, expired, cancelled, past_due] }
        matches_used: { type: integer }
        matches_limit: { type: integer }
        current_period_end: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    AdminSubscriptionList:
      type: object
      properties:
        breakdown: { $ref: '#/components/schemas/AdminTierBreakdown' }
        subscriptions: { type: array, items: { $ref: '#/components/schemas/AdminSubscriptionRow' } }
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
        pages: { type: integer }
    AdminSubscriptionUpdate:
      type: object
      required: [tier]
      properties:
        tier: { type: string, enum: [free, starter, professional, super_standard] }
    Job:
      type: object
      description: |
        Canonical job response shape. Mirrors apps/backend/app/schemas/jobs.py::Job.
        All task-#60 fields are optional + nullable so legacy rows (pre-migration 016)
        still satisfy the schema.
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        company: { type: string }
        location: { type: string }
        description: { type: string }
        requirements: { type: array, items: { type: string } }
        skills_required: { type: array, items: { type: string } }
        skills: { type: array, items: { type: string }, description: "Alias for frontend compatibility" }
        salary_min: { type: integer }
        salary_max: { type: integer }
        apply_url: { type: string, format: uri }
        apply_email: { type: string, format: email }
        source: { type: string, enum: [manual, scraper, ocr, partner] }
        source_url: { type: string, format: uri }
        quality_score: { type: integer, minimum: 0, maximum: 100 }
        closing_date: { type: string, format: date }
        posted_at: { type: string, format: date-time }
        is_active: { type: boolean }
        # task #60: richer job ad shape (migration 016). All nullable.
        employment_type:
          type: string
          enum: [full_time, part_time, contract, freelance, internship, temporary]
          nullable: true
        work_arrangement:
          type: string
          enum: [remote, hybrid, on_site]
          nullable: true
        hybrid_days_per_week: { type: integer, minimum: 1, maximum: 5, nullable: true }
        benefits: { type: array, items: { type: string }, nullable: true }
        application_instructions: { type: string, nullable: true }
        reporting_structure: { type: string, nullable: true }
        manages_others: { type: integer, minimum: 0, nullable: true }
        interview_process: { type: string, nullable: true }
        tools_tech_stack: { type: array, items: { type: string }, nullable: true }
        success_metrics: { type: string, nullable: true }
        company_description: { type: string, nullable: true }
        reference_number: { type: string, nullable: true }
        currency: { type: string, nullable: true }
        pay_frequency:
          type: string
          enum: [monthly, annual, hourly, daily]
          nullable: true
        bonus_structure: { type: string, nullable: true }
        equity_offered: { type: boolean, nullable: true }
    JobCreate:
      type: object
      required: [title, description, source]
      properties:
        title: { type: string, minLength: 5 }
        company: { type: string }
        location: { type: string }
        description: { type: string, minLength: 20 }
        requirements: { type: array, items: { type: string } }
        skills_required: { type: array, items: { type: string } }
        salary_min: { type: integer }
        salary_max: { type: integer }
        apply_url: { type: string, format: uri }
        apply_email: { type: string, format: email }
        source: { type: string, enum: [manual, scraper, ocr, partner] }
        closing_date: { type: string, format: date }
...
    JobList:
      type: object
      properties:
        jobs: { type: array, items: { $ref: '#/components/schemas/Job' } }
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
        pages: { type: integer }
    MatchResult:
      type: object
      properties:
        id: { type: string, format: uuid }
        job: { $ref: '#/components/schemas/Job' }
        score: { type: number, minimum: 0, maximum: 100 }
        vector_score: { type: number }
        skill_score: { type: number }
        bonus_score: { type: number }
        matched_skills: { type: array, items: { type: string } }
        missing_skills: { type: array, items: { type: string } }
        explanation: { type: string }
        created_at: { type: string, format: date-time }
    MatchList:
      type: object
      properties:
        matches: { type: array, items: { $ref: '#/components/schemas/MatchResult' } }
        remaining_quota: { type: integer }
        credited_count: { type: integer }
        matches_limit: { type: integer }
    CronTickResponse:
      type: object
      properties:
        users_processed: { type: integer }
        new_matches_total: { type: integer }
    NotificationDigestResponse:
      type: object
      properties:
        users_processed: { type: integer }
        notifications_sent: { type: integer }
    Subscription:
      type: object
      description: |
        Shape of GET /subscription. Mirrors apps/backend/app/schemas/subscription.py::Subscription
        (the API response model — distinct from the wider DB row, which carries id,
        status, and current_period_*). `active` is true when status='active'.
      properties:
        tier: { type: string, enum: [free, starter, professional, super_standard] }
        matches_used: { type: integer }
        matches_limit: { type: integer }
        active: { type: boolean }
        expires_at: { type: string, format: date-time, nullable: true }
    PaymentInitiate:
      type: object
      required: [tier, payment_method]
      properties:
        tier: { type: string, enum: [starter, professional, super_standard] }
        payment_method: { type: string, enum: [mtn_money, airtel_money] }
        phone: { type: string, pattern: '^\+260[0-9]{9}$' }
...
    PaymentInitiateResponse:
      type: object
      properties:
        transaction_token: { type: string }
        payment_url: { type: string, format: uri }
        status: { type: string, enum: [pending, redirect] }
...
paths:
...
  /admin/matches:
    get:
      operationId: adminListMatches
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, default: 25 } }
        - { name: min_score, in: query, schema: { type: number, minimum: 0, maximum: 100 } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminMatchList' }
        '403': { description: Superadmin only }
  /admin/subscriptions:
    get:
      operationId: adminListSubscriptions
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, default: 25 } }
        - { name: tier, in: query, schema: { type: string, enum: [free, starter, professional, super_standard] } }
        - { name: status, in: query, schema: { type: string, enum: [active, expired, cancelled, past_due] } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminSubscriptionList' }
        '403': { description: Superadmin only }
  /admin/subscriptions/{user_id}:
    parameters:
      - { in: path, name: user_id, required: true, schema: { type: string, format: uuid } }
    patch:
      operationId: adminUpdateSubscription
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminSubscriptionUpdate' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminSubscriptionRow' }
        '403': { description: Superadmin only }
        '404': { description: Subscription not found }
...
  /jobs:
    get:
      operationId: listJobs
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, default: 20 } }
        - { name: location, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobList' }
    post:
      operationId: createJob
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JobCreate' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
...
  /matches:
    get:
      operationId: getMatches
      tags: [Matching]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: min_score, in: query, schema: { type: number, default: 50 } }
        - { name: limit, in: query, schema: { type: integer, default: 10 } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MatchList' }
  /matches/trigger:
    post:
      operationId: triggerMatching
      tags: [Matching]
      security: [{ bearerAuth: [] }]
      responses:
        '202': { description: Matching queued }
  /matches/cron-tick:
    post:
      operationId: matchCronTick
      tags: [Matching]
      parameters:
        - { name: INGEST_API_KEY, in: header, required: true, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CronTickResponse' }
        '401': { description: Invalid ingest API key }
  /matches/send-notifications:
    post:
      operationId: sendMatchNotifications
      tags: [Matching]
      parameters:
        - { name: INGEST_API_KEY, in: header, required: true, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationDigestResponse' }
        '401': { description: Invalid ingest API key }
...
  /subscription:
    get:
      operationId: getSubscription
      tags: [Subscription]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscription' }
  /subscription/pay:
    post:
      operationId: initiatePayment
      tags: [Subscription]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaymentInitiate' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentInitiateResponse' }
...